Added HP bar (moves and doesn't work)

+++ core/src/com/mygdx/game/Player.java
@@ -25,9 +25,9 @@ public class Player extends Sprite {
     float v;
     public boolean isWinner = false;
     static Sound bitShoot;
-    public float hp;
+    public int hp;
     public boolean dead;
-    public Player(Texture img, float hp, float width, float height, float x, float y, float moveSpeed, float powerJump, float gravity) {
+    public Player(Texture img, int hp, float width, float height, float x, float y, float moveSpeed, float powerJump, float gravity) {
         super(img, 0, 0, img.getWidth(), img.getHeight());
         this.x = x;
         this.y = y;
